Remove vconf ipc based logic for syncronizing with master
We now have systemd auto activation support for whatever process
provides the notification socket.  As a result we can now greatly
simplify the client logic that was attempting to provide the same
capability as socket activation in that the client was able to
register with the server before the server was actually started.

Change-Id: I93a8af60cfba30f775cfa0af90c8c267c9678819
Signed-off-by: Rusty Lynch <rusty.lynch@intel.com>