Implement a notification response wait/send API
Currently, interactive responses for notifications (done in
notification-display-service) are sent using a framework-
dependent mechanism. This is prone to become unmaintainable
as the scope of GUI notifications broadens.

There is a need for a generic way to wait for notifications
responses on one side, and to send such a response on
another side. We introduce 2 new functions here :

- notification_wait_response() : retrieves an integer and
a char* from the response, with an optional timeout value.

- notification_send_response() : send a response under the
form of an integer and an optional char*.

Communication is done with sockets using a TID (thread ID)
as an identifier.

Bug-Tizen: TC-996

Change-Id: I86d9d4bc3399dddff03b7f6c7eacbe2f842c5019
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>